requirements.txt
# Server Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
bcrypt==4.1.1
python-dotenv==1.0.0

# Client Dependencies
pycryptodome==3.19.0
requests==2.31.0
cryptography==41.0.7
PyYAML==6.0.1

# Optional Features
gitpython==3.1.40
qrcode==7.4.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0
